requirements.txt
django>=4.0
djangorestframework
psycopg2-binary
celery
django-celery-results
openpyxl
redis